// frontend/projects/ng-advanced-grid/src/lib/interfaces/grid.interfaces.ts
export interface GridPosition {
  x: number;
  y: number;
}

export interface GridSize {
  width: number;
  height: number;
}

export interface GridBounds {
  top: number;
  left: number;
  right: number;
  bottom: number;
}

export interface GridItem {
  id: string;
  x: number;
  y: number;
  width: number;
  height: number;
  minWidth?: number;
  minHeight?: number;
  maxWidth?: number;
  maxHeight?: number;
  fixed?: boolean;
  resizable?: boolean;
  draggable?: boolean;
  data?: any;
}
// ...
export interface GridBreakpoint {
  breakpoint: string;
  minWidth: number;
  columns: number;
  margin?: number;
  cellHeight?: number;
}
// ...
export interface CollisionResult {
  hasCollision: boolean;
  collisions: GridItem[];
  suggestedPosition?: GridPosition;
}